replace func code in presenter from webCont and delete there
