+++ ImageFeed/Auth/WebViewPresenter.swift
@@ -1,9 +1,11 @@
 import Foundation
+import WebKit
 
 public protocol WebViewPresenterProtocol {
     var view: WebViewControllerProtocol? { get set }
     func loadAuthView()
     func didUpdateProgressValue(_ newValue: Double)
+    func code(from navigationAction: WKNavigationAction) -> String?
 }
 
 final class WebViewPresenter: WebViewPresenterProtocol {
@@ -43,6 +45,20 @@ final class WebViewPresenter: WebViewPresenterProtocol {
     
     func shouldHideProgress(for value: Float) -> Bool {
         abs(value - 1.0) <= 0.0001
+    }
+    
+    func code(from navigationAction: WKNavigationAction) -> String? {
+        if
+            let url = navigationAction.request.url,
+            let urlComponents = URLComponents(string: url.absoluteString), 
+            urlComponents.path == "/oauth/authorize/native",
+            let items = urlComponents.queryItems,
+            let codeItem = items.first(where: { $0.name == "code" })
+        {
+            return codeItem.value
+        } else {
+            return nil
+        }
     }
     
 }
